+++ CHANGES.rst
@@ -1,7 +1,7 @@
 0.14.0 (unreleased)
 ===================
 
--
+- Explicitly add ``gwcs`` to the list of dependencies. [#108]
 
 0.13.0 (2022-08-23)
 ===================