+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ===================
 
 - Explicitly add ``gwcs`` to the list of dependencies. [#108]
+- Add support for non-VOUnits to be used by Roman. [#109]
 
 0.13.0 (2022-08-23)
 ===================